agents_server/requirements.txt
faiss-cpu
pandas
numpy
sentence-transformers
scikit-learn
neo4j
tqdm
python-dotenv
google-generativeai
requests
chromadb
pyautogen
openai
motor
tenacity
pydantic
fastapi
uvicorn
